requirements.txt
# Construction AI Agent dependencies

# Core AI & orchestration
openai>=1.58
langchain>=0.3
langchain-openai>=0.2
langchain-community>=0.3

duckduckgo-search>=6.3

# Google Sheets integration
gspread>=6.1
google-auth>=2.37
google-auth-oauthlib>=1.2

# Web scraping & parsing
requests>=2.32
beautifulsoup4>=4.12
pydantic>=2.9
tenacity>=9.0
tabulate>=0.9
python-dotenv>=1.0

# Web application
flask>=3.0
flask-cors>=4.0
waitress>=2.1
